requirements.txt
streamlit
pandas
google-generativeai
pydantic
python-dotenv
openpyxl
xlsxwriter